Add endpoint for retrieving discard remark suggestions
This endpoint can be used to retrieve suggestions for the instance
discard offcanvas input. Supports case-insensitive any-position matching
through the `name` parameter. Results are sorted according to the start
position of the match, followed by an alphabetical sort. This endpoint
is also paginated by default in order to control the number of returned
results.